SequenceViewer fixes
git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@426 6721810a-1124-0410-938e-bb277542c816

+++ jebl/src/jebl/evolution/align/ProfileCharacter.java
@@ -90,7 +90,7 @@ public class ProfileCharacter {
 
     public static float scoreSelf(ProfileCharacter character, Scores scores) {
         float score = 0;
-        int totalCharacters = character.totalCharacters * character.totalCharacters;
+        long totalCharacters = ((long)character.totalCharacters) * character.totalCharacters;
         if (totalCharacters == 1) {
             return scores.score[character.characters[0]][character.characters[0]];
         }
